renames current turn to board
