+++ game_lib/src/world/deck.rs
@@ -11,7 +11,7 @@ use crate::cards::types::oopsie::OopsieCard;
 
 #[derive(Debug, Clone)]
 pub struct Deck {
-    pub cards: Vec<Card>,
+    pub board: Vec<Card>,
     pub played_cards: usize,
     pub total: usize,
 }
@@ -20,7 +20,7 @@ impl Deck {
     fn new(cards: Vec<Card>) -> Deck {
         let total = cards.len();
         Deck {
-            cards,
+            board: cards,
             played_cards: 0,
             total,
         }
@@ -28,7 +28,7 @@ impl Deck {
 
     pub(crate) fn with_remaining_cards(&self, cards: &[Card]) -> Self {
         Deck {
-            cards: cards.to_vec(),
+            board: cards.to_vec(),
             played_cards: self.played_cards + 1,
             total: self.total,
         }
